Moved logits `.float()` to loss and compiled it if compiling
[ghstack-poisoned]

+++ torchtitan/models/llama/model.py
@@ -439,7 +439,7 @@ class Transformer(nn.Module):
             h = layer(h, self.freqs_cis)
 
         h = self.norm(h) if self.norm else h
-        output = self.output(h).float() if self.output else h
+        output = self.output(h) if self.output else h
         return output
 
     @classmethod
